update package name of test classes
