+++ flink-connector-oceanbase-tools-cdc/src/test/java/com/oceanbase/connector/flink/tools/DatabaseSyncTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.oceanbase.cdc.tools.tests.tools;
+package com.oceanbase.connector.flink.tools;
 
 import com.oceanbase.connector.flink.tools.cdc.DatabaseSync;
 import com.oceanbase.connector.flink.tools.cdc.db2.Db2DatabaseSync;
